Added widgets for pets_screen
